// CoinCollectionFrontend/src/api/models/CoinGroupByCurrencyDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CoinGroupByCurrencyItemDto } from './CoinGroupByCurrencyItemDto';
import type { CurrencyDto } from './CurrencyDto';
export type CoinGroupByCurrencyDto = {
    currency?: CurrencyDto;
    items?: Array<CoinGroupByCurrencyItemDto> | null;
};